moved logic out
