+++ src/actions/actions.js
@@ -1,9 +1,12 @@
+import mongodb from 'mongodb';
+import multer from 'multer';
+import { Readable } from 'stream';
 import admin from './firebase';
 import translate from './translate';
 import speech from './speech';
 
 const database = admin.database();
-const bucket = admin.storage().bucket('learn-igbo');
+const ObjectID = mongodb.ObjectID;
 
 const translateInput = (req, res) => {
   console.log(req.headers);
@@ -25,6 +28,72 @@ const translateInput = (req, res) => {
       errorMessage: error.message
     });
   })
+}
+
+const uploadAudio = (req, res, database) => {
+  const storage = multer.memoryStorage();
+  const upload = multer({ storage, limits: { fields: 1, fileSize: 6000000, files: 1, parts: 2 } });
+
+  upload.single('track')(req, res, (err) => {
+    if (err) {
+      return res.status(400).json({ message: "Upload Request Validation Failed" });
+    } else if (!req.body.name) {
+      return res.status(400).json({ message: "No track name in request body" });
+    }
+
+    let trackName = req.body.name;
+
+    // Covert buffer to Readable Stream
+    const readableTrackStream = new Readable();
+    readableTrackStream.push(req.file.buffer);
+    readableTrackStream.push(null);
+
+    let bucket = new mongodb.GridFSBucket(database, {
+      bucketName: 'tracks'
+    });
+
+    let uploadStream = bucket.openUploadStream(trackName);
+    let id = uploadStream.id;
+    readableTrackStream.pipe(uploadStream);
+
+    uploadStream.on('error', () => {
+      return res.status(500).json({ message: "Error uploading file" });
+    });
+
+    uploadStream.on('finish', () => {
+      return res.status(201).json({ message: "File uploaded successfully, stored under Mongo ObjectID: " + id });
+    });
+
+  })
+}
+
+const downloadAudio = (req, res, database) => {
+  try {
+    var id = new ObjectID(req.params.id);
+  } catch (err) {
+    console.log(err);
+    return res.status(400).json({ message: "Invalid id in URL parameter. Must be a single String of 12 bytes or a string of 24 hex characters" });
+  }
+  res.set('content-type', 'audio/mp3');
+  res.set('accept-ranges', 'bytes');
+
+  let bucket = new mongodb.GridFSBucket(database, {
+    bucketName: 'tracks'
+  });
+
+  let downloadStream = bucket.openDownloadStream(id);
+
+  downloadStream.on('data', (chunk) => {
+    res.write(chunk);
+  });
+
+  downloadStream.on('error', () => {
+    res.sendStatus(404);
+  });
+
+  downloadStream.on('end', () => {
+    res.end();
+  });
 }
 
 const convertSpeech = (req, res) => {
@@ -73,4 +142,4 @@ const convertSpeech = (req, res) => {
 
 }
 
-export default { translateInput, convertSpeech };
+export default { translateInput, convertSpeech, uploadAudio, downloadAudio };
